Update quantity columns to float and fix variance calculation
Changed quantity-related columns in several tables from integer to float for improved precision via a new migration. Updated variance calculation in StoreReceivingReportController to use float values instead of integers.

+++ app/Http/Controllers/v1/Report/StoreReceivingReportController.php
@@ -90,7 +90,7 @@ class StoreReceivingReportController extends Controller
                 $receivedAt = $item['formatted_received_at_label'] ?? null;
                 $deliveryDate = $item['formatted_delivery_date_label'] ?? null;
                 $remarks = null;
-                $variance = intval($receivedQuantity) - intval($allocatedQuantity);
+                $variance = floatval($receivedQuantity) - floatval($allocatedQuantity);
 
                 if ($receivedQuantity < $allocatedQuantity) {
                     $remarks = 'Short';
